--- mobile/src/navigation/screens/Album/AlbumsScreen.tsx
@@ -1,15 +1,0 @@
-import { useAlbumsForCards } from "~/queries/album";
-import { StickyActionListLayout } from "~/layouts/StickyActionScroll";
-
-import { useMediaCardListPreset } from "~/modules/media/components/MediaCard";
-
-export default function AlbumsScreen() {
-  const { isPending, data } = useAlbumsForCards();
-  const presets = useMediaCardListPreset({
-    data,
-    isPending,
-    errMsgKey: "err.msg.noAlbums",
-  });
-
-  return <StickyActionListLayout titleKey="term.albums" {...presets} />;
-}
