--- mobile/src/navigation/screens/Setting/SaveErrorsScreen.tsx
@@ -1,25 +1,0 @@
-import { useSaveErrors } from "~/queries/setting";
-
-import { useListPresets } from "~/components/Containment/List";
-import { FlashList } from "~/components/Defaults";
-import { ContentPlaceholder } from "~/components/Transition/Placeholder";
-
-export default function SaveErrorsScreen() {
-  const { data } = useSaveErrors();
-  const presets = useListPresets({
-    data,
-    renderOptions: {
-      getTitle: (item) => item.uri,
-      getDescription: (item) => `[${item.errorName}] ${item.errorMessage}`,
-    },
-  });
-
-  return (
-    <FlashList
-      keyExtractor={({ id }) => id}
-      ListEmptyComponent={<ContentPlaceholder errMsgKey="err.msg.noErrors" />}
-      contentContainerClassName="p-4"
-      {...presets}
-    />
-  );
-}
