+++ mobile/src/navigation/screens/tracks/ModifyView.tsx
@@ -1,0 +1,14 @@
+import { useLocalSearchParams } from "expo-router";
+
+import { useTrack } from "~/queries/track";
+
+import { ModifyTrack as ModifyTrackBase } from "~/screens/ModifyTrack";
+import { PagePlaceholder } from "~/components/Transition/Placeholder";
+
+export default function ModifyTrack() {
+  const { id } = useLocalSearchParams<{ id: string }>();
+  const { isPending, error, data } = useTrack(id);
+
+  if (isPending || error) return <PagePlaceholder isPending={isPending} />;
+  return <ModifyTrackBase initialData={data} />;
+}
